+++ src/AdminMenus/ReusableBlocksHeaderFooterCli.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Class that registers WPCLI command for Admin reusable blocks menu creation.
+ * Class that registers WPCLI command for Reusable header/footer picker creation.
  *
  * @package EightshiftLibs\AdminMenus
  */
@@ -15,9 +15,9 @@ use EightshiftLibs\Cli\ParentGroups\CliCreate;
 use EightshiftLibs\Helpers\Components;
 
 /**
- * Class AdminReusableBlocksMenuCli
+ * Class ReusableBlocksHeaderFooterCli
  */
-class AdminReusableBlocksMenuCli extends AbstractCli
+class ReusableBlocksHeaderFooterCli extends AbstractCli
 {
 	/**
 	 * Get WPCLI command parent name
@@ -36,7 +36,7 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 	 */
 	public function getCommandName(): string
 	{
-		return 'admin-reusable-blocks-menu';
+		return 'reusable-header-footer';
 	}
 
 	/**
@@ -47,11 +47,10 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 	public function getDefaultArgs(): array
 	{
 		return [
-			'title' => 'Reusable blocks',
-			'menu_title' => 'Reusable blocks',
+			'title' => 'Header & footer',
+			'menu_title' => 'Header & footer',
 			'capability' => 'edit_posts',
-			'menu_icon' => '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24"><path d="m18.492 2.33 3.179 3.179a2.25 2.25 0 0 1 0 3.182l-2.423 2.422A2.501 2.501 0 0 1 21 13.5v5a2.5 2.5 0 0 1-2.5 2.5h-13A2.5 2.5 0 0 1 3 18.5v-13A2.5 2.5 0 0 1 5.5 3h5a2.5 2.5 0 0 1 2.387 1.754L15.31 2.33a2.25 2.25 0 0 1 3.182 0ZM11 13H5v5.5a.5.5 0 0 0 .5.5H11v-6Zm7.5 0H13v6h5.5a.5.5 0 0 0 .5-.5v-5a.5.5 0 0 0-.5-.5Zm-4.06-2.001L13 9.559v1.44h1.44Zm-3.94-6h-5a.5.5 0 0 0-.5.5V11h6V5.5a.5.5 0 0 0-.5-.5Z" fill="black"/></svg>', // phpcs:ignore
-			'menu_position' => 4,
+			'menu_position' => 59,
 		];
 	}
 
@@ -63,7 +62,7 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 	public function getDoc(): array
 	{
 		return [
-			'shortdesc' => 'Create reusable blocks admin menu service class.',
+			'shortdesc' => 'Create reusable blocks header/footer admin menu service class.',
 			'synopsis' => [
 				[
 					'type' => 'assoc',
@@ -88,13 +87,6 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 				],
 				[
 					'type' => 'assoc',
-					'name' => 'menu_icon',
-					'description' => 'The default menu icon for the admin menu.',
-					'optional' => true,
-					'default' => $this->getDefaultArg('menu_icon'),
-				],
-				[
-					'type' => 'assoc',
 					'name' => 'menu_position',
 					'description' => 'The default menu position.',
 					'optional' => true,
@@ -104,17 +96,16 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 			'longdesc' => $this->prepareLongDesc("
 				## USAGE
 
-				Used to create top level sidebar menu page for reusable blocks for easy usage.
+				Used to initialize reusable block header/footer - adds pickers to the admin.
 
 				## EXAMPLES
 
-				# Create service class:
 				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()}
 
 				## RESOURCES
 
 				Service class will be created from this example:
-				https://github.com/infinum/eightshift-libs/blob/develop/src/AdminMenus/AdminReusableBlocksMenuExample.php
+				https://github.com/infinum/eightshift-libs/blob/develop/src/AdminMenus/ReusableBlocksHeaderFooterExample.php
 			"),
 		];
 	}
@@ -127,9 +118,8 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 		// Get Arguments.
 		$title = $this->getArg($assocArgs, 'title');
 		$menuTitle = $this->getArg($assocArgs, 'menu_title');
-		$capability = $this->getArg($assocArgs, 'capability');
-		$menuIcon = $this->getArg($assocArgs, 'menu_icon');
 		$menuPosition = (string)($this->getArg($assocArgs, 'menu_position'));
+		$capability = $this->getArg($assocArgs, 'capability');
 
 		// Get full class name.
 		$className = $this->getFileName('');
@@ -142,15 +132,11 @@ class AdminReusableBlocksMenuCli extends AbstractCli
 			->renameUse($assocArgs)
 			->renameTextDomain($assocArgs)
 			->searchReplaceString($this->getArgTemplate('title'), $title)
-			->searchReplaceString($this->getArgTemplate('menu_title'), $menuTitle)
-			->searchReplaceString($this->getArgTemplate('capability'), $capability);
+			->searchReplaceString($this->getArgTemplate('capability'), $capability)
+			->searchReplaceString($this->getArgTemplate('menu_title'), $menuTitle);
 
 		if (!empty($menuPosition)) {
 			$class->searchReplaceString($this->getDefaultArg('menu_position'), $menuPosition);
-		}
-
-		if (!empty($menuIcon)) {
-			$class->searchReplaceString($this->getArgTemplate('menu_icon'), $menuIcon);
 		}
 
 		// Output final class to new file/folder and finish.