+++ src/Columns/Media/AbstractMediaColumns.php
@@ -26,9 +26,9 @@ abstract class AbstractMediaColumns implements ServiceInterface
 	 */
 	public function register(): void
 	{
-		\add_filter('manage_media_columns', [$this, 'addColumnName']);
+		\add_filter('manage_upload_columns', [$this, 'addColumnName']);
 		\add_filter('manage_media_custom_column', [$this, 'renderColumnContent'], 10, 2);
-		\add_filter('manage_media_sortable_columns', [$this, 'sortAddedColumns'], 10);
+		\add_filter('manage_upload_sortable_columns', [$this, 'sortAddedColumns'], 10);
 	}
 
 	/**