+++ CHANGELOG.md
@@ -3,10 +3,43 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [6.5.5] - 2023-07-07
+
+### Fixed
+- if you're using the WPML plugin, webP images won't be deleted unless all the attachments are removed from the admin interface.
+
+## [6.5.4] - 2023-07-04
+
+### Fixed
+- you no longer need to add the "blockSsr" attribute when loading a block via server-side rendering. It will automatically utilize global CSS variables and provide the necessary styles.
+
+## [6.5.3] - 2023-06-29
+
+### Fixed
+- column media wrong filter name.
+- deprecation notice in the CLI.
+
+## [6.5.2] - 2023-06-20
+
+### Changed
+- updating random Id generator to output 8 rather than 32 characters.
+
+## [6.5.1] - 2023-06-19
+
+### Changed
+- wrong command name was changed from `plugin_manage` to `plugin-manage`.
+- webp helper now has option to add allowed item so you can limit image types you don't want to use webP.
+
+## [6.5.0] - 2023-06-13
+
+### Changed
+- added new CLI command for initializing reusable header/footer (settings page for picking blocks, and default blocks)
+- added new helper to render nice alert "windows"
+- revamped some CLI command outputs and error outputs
+
 ## [6.4.0] - 2023-02-21
 
 ### Changed
-
 - Changed the default args for init blocks and use block variation CLI commands, as the button-block variation will not exist anymore in FE libs v8.0.
 
 ## [6.3.2] - 2023-01-27
@@ -401,6 +434,12 @@ Init setup
 
 [Unreleased]: https://github.com/infinum/eightshift-libs/compare/main...HEAD
 
+[6.5.5]: https://github.com/infinum/eightshift-libs/compare/6.5.4...v6.5.5
+[6.5.4]: https://github.com/infinum/eightshift-libs/compare/6.5.3...v6.5.4
+[6.5.3]: https://github.com/infinum/eightshift-libs/compare/6.5.2...v6.5.3
+[6.5.2]: https://github.com/infinum/eightshift-libs/compare/6.5.1...v6.5.2
+[6.5.1]: https://github.com/infinum/eightshift-libs/compare/6.5.0...v6.5.1
+[6.5.0]: https://github.com/infinum/eightshift-libs/compare/6.4.0...v6.5.0
 [6.4.0]: https://github.com/infinum/eightshift-libs/compare/6.3.2...v6.4.0
 [6.3.2]: https://github.com/infinum/eightshift-libs/compare/6.3.1...v6.3.2
 [6.3.1]: https://github.com/infinum/eightshift-libs/compare/6.3.0...v6.3.1